Update enums and namespace

+++ src/ToolingSharedLibrary/CurrentCodeGenerationState/Rust/HostApp/code_gen_test_gen/src/implementation/types.rs
@@ -9,10 +9,9 @@ use crate::abi::fb_support::fb_types::code_gen_test::flatbuffer_types;
 pub use crate::abi::abi_types::edl;
 use edlcodegen_host::EdlDerive;
 
-pub const VALUE1: u32 = 1;
-
+pub const VALUE1: usize = 1;
 #[repr(u32)]
-#[derive(Debug, Clone, PartialEq, Default, EdlDerive)]
+#[derive(Debug, Copy, Clone, PartialEq, Default, EdlDerive)]
 #[target_enum(flatbuffer_types::TestEnum)]
 pub enum TestEnum {
     #[default]
